// src/codegen/ArenaPaymentRegistry.types.ts
/**
* This file was automatically generated by @cosmwasm/ts-codegen@1.12.0.
* DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
* and run the @cosmwasm/ts-codegen generate command to regenerate this file.
*/

export interface InstantiateMsg {}
export type ExecuteMsg = {
  set_distribution: {
    distribution: DistributionForString;
  };
} | {
  set_distribution_remainder_self: {
    member_percentages: MemberPercentageForString[];
  };
} | {
  remove_distribution: {};
};
export type Decimal = string;
export interface DistributionForString {
  member_percentages: MemberPercentageForString[];
  remainder_addr: string;
}
export interface MemberPercentageForString {
  addr: string;
  percentage: Decimal;
}
export type QueryMsg = {
  get_distribution: {
    addr: string;
    height?: number | null;
  };
};
export type MigrateMsg = {
  from_compatible: {};
};
export type NullableDistributionForString = DistributionForString | null;